Reader: parse escapes in strings
All escapes described in R7RS are supported.

+++ sexp/uchar.h
@@ -100,17 +100,17 @@ size_t u_strlen_safe(const char *str);
  * Returns unicode character at @str[*@idx] or @str[0] if @idx is NULL.
  * Stores byte index of the next char back to @idx if set.
  */
-uchar u_get_char(const char *str, unsigned *idx);
+uchar u_get_char(const char *str, size_t *idx);
 
 /*
  * @str  destination buffer
  * @idx  pointer to byte index in @str (not UTF-8 character index!)
  * @uch  unicode character
  */
-void u_set_char_raw(char *str, unsigned *idx, uchar uch);
-void u_set_char(char *str, unsigned *idx, uchar uch);
+void u_set_char_raw(char *str, size_t *idx, uchar uch);
+void u_set_char(char *str, size_t *idx, uchar uch);
 
-static inline void u_skip_chars(const char *str, int n, unsigned *idx)
+static inline void u_skip_chars(const char *str, int n, size_t *idx)
 {
 	while (n-- > 0)
 		u_get_char(str, idx);
