+++ uchar.c
@@ -167,11 +167,11 @@ single_char:
 	return len;
 }
 
-uchar u_get_char(const char *str, unsigned *idx)
+uchar u_get_char(const char *str, size_t *idx)
 {
 	const unsigned char *s = (const unsigned char *)str;
 	int len, i;
-	unsigned x = 0;
+	size_t x = 0;
 	uchar ch, u;
 
 	if (idx)
@@ -205,9 +205,9 @@ invalid:
 	return u | U_INVALID_MASK;
 }
 
-void u_set_char_raw(char *str, unsigned *idx, uchar uch)
+void u_set_char_raw(char *str, size_t *idx, uchar uch)
 {
-	int i = *idx;
+	size_t i = *idx;
 
 	if (uch <= 0x0000007fU) {
 		str[i++] = uch;
@@ -241,9 +241,9 @@ void u_set_char_raw(char *str, unsigned *idx, uchar uch)
  * Printing functions, these lose information
  */
 
-void u_set_char(char *str, unsigned *idx, uchar uch)
+void u_set_char(char *str, size_t *idx, uchar uch)
 {
-	int i = *idx;
+	size_t i = *idx;
 
 	if (unlikely(uch <= 0x0000001fU))
 		goto invalid;
